+++ src/effects/specialMonsters.js
@@ -33,7 +33,6 @@ export function monsterFeatEffectModules() {
   if (installedModules) return installedModules;
   const midiQolInstalled = game.modules.get("midi-qol")?.active;
   const advancedMacrosInstalled = game.modules.get("advanced-macros")?.active;
-  const aboutTime = game.modules.get("about-time")?.active;
   const itemMacroInstalled = game.modules.get("itemacro")?.active;
   const timesUp = game.modules.get("times-up")?.active;
   const daeInstalled = game.modules.get("dae")?.active;
@@ -49,14 +48,12 @@ export function monsterFeatEffectModules() {
       itemMacroInstalled &&
       midiQolInstalled &&
       advancedMacrosInstalled &&
-      aboutTime &&
       timesUp &&
       daeInstalled &&
       convenientEffectsInstalled,
     midiQolInstalled,
     itemMacroInstalled,
     advancedMacrosInstalled,
-    aboutTime,
     timesUp,
     daeInstalled,
     convenientEffectsInstalled,